Move EmailSender to Mail namespace.
